src/main/java/com/waracle/cakemgr/service/CakeService.java
package com.waracle.cakemgr.service;

import com.waracle.cakemgr.model.Cake;
import com.waracle.cakemgr.repository.CakeRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class CakeService {

    @Autowired
    private CakeRepository cakeRepository;

    public List<Cake> getCakes() {
        return cakeRepository.getCakes();
    }

    public Cake addCake(Cake cake) {
        return cakeRepository.saveCake(cake);
    }
}
